add initial patient model
