+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/BackboneElement.java
@@ -1,0 +1,7 @@
+package gov.va.api.health.argonaut.api;
+
+import java.util.List;
+
+public interface BackboneElement extends Element{
+    List<Extension> modifierExtension();
+}
