+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/ContactPoint.java
@@ -1,0 +1,35 @@
+package gov.va.api.health.argonaut.api;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import lombok.Builder;
+import lombok.Data;
+
+import javax.validation.Valid;
+import javax.validation.constraints.Min;
+import javax.validation.constraints.Pattern;
+import java.util.List;
+
+@Data
+@Builder
+@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+public class ContactPoint implements Element {
+    @Pattern(regexp = Fhir.ID)
+    String id;
+
+    @Valid List<Extension> extension;
+    ContactPointSystem system;
+    String value;
+
+    @Min(1)
+    Integer rank;
+
+    @Valid Period period;
+
+    enum ContactPointSystem {
+        phone,
+        fax,
+        email,
+        pager,
+        other
+    }
+}
