+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/Element.java
@@ -1,0 +1,9 @@
+package gov.va.api.health.argonaut.api;
+
+import java.util.List;
+
+public interface Element {
+    String id();
+
+    List<Extension> extension();
+}
