+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/Fhir.java
@@ -28,345 +28,4 @@ public class Fhir {
   public static final String OID = "urn:oid:[0-2](\\.[1-9]\\d*)+";
   public static final String XHTML = "<.+>";
 
-  public interface Element {
-    String id();
-
-    List<Extension> extension();
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class Address implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    AddressUse use;
-    AddressType type;
-    String text;
-    List<String> line;
-    String city;
-    String district;
-    String state;
-    String postalCode;
-    String country;
-    @Valid Period period;
-
-    public enum AddressUse {
-      home,
-      work,
-      temp,
-      old
-    }
-
-    private enum AddressType {
-      postal,
-      physical,
-      both
-    }
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class Attachment implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    @Pattern(regexp = CODE)
-    String contentType;
-
-    @Pattern(regexp = CODE)
-    String language;
-
-    @Pattern(regexp = BASE64)
-    String data;
-
-    @Pattern(regexp = URI)
-    String url;
-
-    @Min(0)
-    Integer size;
-
-    @Pattern(regexp = BASE64)
-    String hash;
-
-    String title;
-
-    @Pattern(regexp = DATETIME)
-    String creation;
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class CodeableConcept implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    @Valid List<Coding> coding;
-    String text;
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class Coding {
-    @Pattern(regexp = URI)
-    String system;
-
-    String version;
-
-    @Pattern(regexp = CODE)
-    String code;
-
-    String display;
-    Boolean userSelected;
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class ContactPoint implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-    ContactPointSystem system;
-    String value;
-
-    @Min(1)
-    Integer rank;
-
-    @Valid Period period;
-
-    enum ContactPointSystem {
-      phone,
-      fax,
-      email,
-      pager,
-      other
-    }
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class Extension implements Element {
-
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    @Pattern(regexp = URI)
-    String url;
-
-    Integer valueInteger;
-    Double valueDecimal;
-
-    @Pattern(regexp = DATETIME)
-    String valueDateTime;
-
-    @Pattern(regexp = DATE)
-    String valueDate;
-
-    @Pattern(regexp = INSTANT)
-    String valueInstant;
-
-    String valueString;
-
-    @Pattern(regexp = URI)
-    String valueUri;
-
-    Boolean valueBoolean;
-
-    @Pattern(regexp = CODE)
-    String valueCode;
-
-    @Pattern(regexp = BASE64)
-    String valueBase64Binary;
-
-    @Valid Coding valueCoding;
-    @Valid CodeableConcept valueCodeableConcept;
-    @Valid Attachment valueAttachment;
-    @Valid Identifier valueIdentifier;
-    @Valid Quantity valueQuantity;
-    @Valid Range valueRange;
-    @Valid Period valuePeriod;
-    @Valid Ratio valueRatio;
-    @Valid HumanName valueHumanName;
-    @Valid Address valueAddress;
-    @Valid ContactPoint valueContactPoint;
-    @Valid Reference valueReference;
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class HumanName implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    NameUse use;
-    String text;
-    List<String> family;
-    List<String> given;
-    List<String> prefix;
-    List<String> suffix;
-    @Valid Period period;
-
-    public enum NameUse {
-      usual,
-      official,
-      temp,
-      nickname,
-      anonymous,
-      old,
-      maiden
-    }
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class Meta implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    @Pattern(regexp = ID)
-    String versionId;
-
-    @Pattern(regexp = INSTANT)
-    String lastUpdated;
-
-    List<@Pattern(regexp = URI) String> profile;
-    @Valid List<Coding> security;
-    @Valid List<Coding> tag;
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class Narrative implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    @NotNull NarrativeStatus status;
-
-    @NotBlank
-    @Pattern(regexp = XHTML)
-    String div;
-
-    public enum NarrativeStatus {
-      generated,
-      extensions,
-      additional,
-      empty
-    }
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class Period implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    @Pattern(regexp = DATETIME)
-    String start;
-
-    @Pattern(regexp = DATETIME)
-    String end;
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class Quantity implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    Double value;
-
-    @Pattern(regexp = "(<|<=|>=|>)")
-    String comparator;
-
-    String unit;
-
-    @Pattern(regexp = CODE)
-    String code;
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class Range implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    @Valid SimpleQuantity low;
-    @Valid SimpleQuantity high;
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class Ratio implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    @Valid Quantity numerator;
-    @Valid Quantity denominator;
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class Reference implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    String reference;
-    String display;
-  }
-
-  @Data
-  @Builder
-  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-  public static class SimpleQuantity implements Element {
-    @Pattern(regexp = ID)
-    String id;
-
-    @Valid List<Extension> extension;
-
-    Double value;
-
-    String unit;
-
-    @Pattern(regexp = CODE)
-    String code;
-  }
 }
