+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/Identifier.java
@@ -1,10 +1,6 @@
 package gov.va.api.health.argonaut.api;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import gov.va.api.health.argonaut.api.Fhir.CodeableConcept;
-import gov.va.api.health.argonaut.api.Fhir.Extension;
-import gov.va.api.health.argonaut.api.Fhir.Period;
-import gov.va.api.health.argonaut.api.Fhir.Reference;
 import java.util.List;
 import javax.validation.Valid;
 import javax.validation.constraints.Pattern;
@@ -14,7 +10,7 @@ import lombok.Data;
 @Data
 @Builder
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-public class Identifier implements Fhir.Element {
+public class Identifier implements Element {
   @Pattern(regexp = Fhir.ID)
   String id;
