+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/Link.java
@@ -1,0 +1,23 @@
+package gov.va.api.health.argonaut.api;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import lombok.Builder;
+import lombok.Data;
+
+import javax.validation.Valid;
+import javax.validation.constraints.Pattern;
+import java.util.List;
+
+@Data
+@Builder
+@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+public class Link implements BackboneElement {
+    @Pattern(regexp = Fhir.ID)
+    String id;
+
+    @Valid List<Extension> modifierExtension;
+    @Valid List<Extension> extension;
+    @Valid Reference other;
+    @Pattern(regexp = Fhir.CODE)
+    String type;
+}
