+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/Reference.java
@@ -1,0 +1,22 @@
+package gov.va.api.health.argonaut.api;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import lombok.Builder;
+import lombok.Data;
+
+import javax.validation.Valid;
+import javax.validation.constraints.Pattern;
+import java.util.List;
+
+@Data
+@Builder
+@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+public class Reference implements Fhir.Element {
+    @Pattern(regexp = Fhir.ID)
+    String id;
+
+    @Valid List<Fhir.Extension> extension;
+
+    String reference;
+    String display;
+}
